Rename README.md to README2.md
Reflections on Airline Manager Project by Mike Mueller

Initially we started to learn more about how WPF and other new Visual Studio and C# functionality works.  A lot of research via on-line tutorials, Visual Studio literature, and books.  We then broke out the initial classes equally and began coding based on our reasearch.  I started small with the seating chart and the read/write classes, but then realized I needed more knowledge on the overall WPF and ViewModel functionality.  I finally was able to write to a text file after a few days of trial and error, with the debug features (stepinto, step over, etc) within Visual Studio one of the main tools I used to follow the flow of data and the output from the program.  This became a bit inefficient after a while, but in the end gave me the confidence that the program was running as intended.  

Overall a great project to learn new topics and write a lot of code.  It was also very helpful to see my partner's perspective and hearing his approach to solving the problem.  I would still like to see some lectures midstream to just assess everyone's progress to date and to re-calibrate our projects.